Moved ShowFileContextMenu to FileExecutor
